+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * `--ignore-errors` now continues to try to execute all run steps for a kapp instead of aborting the kapp (which helps if e.g. the helm chart has been deleted but Terraform resources weren't cleaned up)
 * YAML loading is now strict - if any keys exist that aren't in our structs an error will be thrown
 * `ws create` now supports `-i/-x` selectors to selectively update kapps
+* Defaults can now be set at the stack level (e.g. to pin helm/kubectl binaries per stack)
 
 ## 0.9.0
 * Rename the 'cache' subcommand to 'workspace' because it wasn't clear that users should actually work inside cache directories (i.e. they're not temporary caches).